Renamed directory structure according to components in /user/profile
