+++ src/app/user/profile/_components/MainSection.tsx
