+++ src/app/user/profile/_components/views/AppliedJobs/AppliedJobsView.tsx
